Added confimation button before deleting a user

+++ frontend/src/screens/UserListScreen.js
@@ -28,7 +28,9 @@ const UserListScreen = () => {
     }, [dispatch, userInfo, navigate, successDelete])
 
     const deleteHandler = (id) => {
-        dispatch(deleteUser(id))
+        if (window.confirm('Are you sure you want to delete this user')) {
+            dispatch(deleteUser(id))
+        }
     }
 
     return (
